[IceTV] Rewrite of the timer state machine.
It turns out that the IceTV API and documentation are misleading and the
design is lacking in areas such as temporal coherence. Following the
suggestions given in the documentation will lead to uncertainty and data
loss. Additionally, the bugs in the IceTV server make it nearly impossible
to correctly implement some of the described workflow.

This implementation attempts to address a number of the issues by mostly
performing operations as they occur and works around a number of current
bugs.