Fix bug #457: Accessing IceTV directly after IceTV installation can cause a crash
Ensure that global variable fetcher is created before the IceTV
popup menu is displayed. An IceTV fetch is not initiated.

Add needsRestart=True to the PluginDescriptor for the WHERE_SESSIONSTART
PluginDescriptor to encourage the user to restart the GUI after installing
the IceTV plugin. This will also run the IceTV config wizard if the
user has not already set up IceTV.