Fix Bug #512: Locally created timers not handled properly in IceTV
[EPGFetcher]

In postTimer(), use data in the JSON body of the IceTV response to
the timer creation request instead of using part of the Location
header field in the response.

The Location header field is no longer sent by the server.

If a response contains an ice_timer_id for the newliy created
local timer, call RecordTimer.saveTimer() to save the updated timer]
to timers.xml.