Fix Issue #528: IceTV can return "No matching service" error when the actual error is timer conflict
[EPGFetcher]

In processTimers(), when not completed and not updated,
set the "No matching service" failure response as a sentinal,
and update the same response when a channel match is successfully made,
or when there is a timer conflict.

The response is only queued once.

This avoids having a "No matching service" failure response sent
when the error is in fact "Timer conflict: ..." and ensures that
only a single response is sent for the timer, whether setting the
timer is successful or fails.

Also add the timer name to the "Failed to update" error response for
a timer.