Fix Bug #603: Gaps in IceTV EPG
[RecordTimer]

Add getEventFromEPGId() method to fetch the event for the timer's
event id (or an event id argument), and use that in setting the
event information when initialising the timer if the event id is
supplied when creating the timer.

[IceTV.API]

Update version string.

[IceTV.plugin]

In EPGFetcher.makeChanShowMap(), use the IceTV eit_id as the event
id in preference to API.showIdToEventId(show["id"])

In EPGFetcher.putTimer() & postTimer() supply the timer's event id
to the timer being sent to IceTV.

In EPGFetcher.processTimers() prefer the event returned by
RecordTimerEntry.getEventFromEPGId() over
RecordTimerEntry.getEventFromEPG() to get a more reliable fetch of
the timer's event id and description. Use the event id as a new
argument to EPGFetcher.updateTimer() since the update timer's eit
must match the timer's eit_id on IceTV. Also supply the event id
to the creation of a new timer.

In EPGFetcher.onTimerChanged(), if the timer already has a valid
ice_timer_id, delete the IceTV-side timer and re-create it, because
IceTV does not permit the eit_id of an existing timer to be changed
(and at this point in the code, there's no simple way to find whether
eit_id has changed).

Add an eit argument to EPGFetcher.updateTimer() in case the eit_id
of the timer on IceTV has changed (Daniel Hall has told me that
that can happen, even though the Beyonwiz side can't change the
eit_id of a timer on the IceTV side).

The changes to use the event id to search the EPG instead of the
event times (using RecordTimerEntry.getEventFromEPGId() instead of
RecordTimerEntry.getEventFromEPG()) helps reduce the incidence of
Bug #174.