Allow batching of EPG fetches to limit memory use
Allow the EPG to be fetched in batches of channels.

The aim is to reduce the maximum memory requirement when downloding
and processing the EPG. The larger (~140 channels, 14 days) German
EPG appeared to be causing OOM problems on some boxes.

config.plugins.icetv.batchsize controls the maximum size of the
batch.  When set to 0, the EPG is loaded in a single fetch as before.
Defaults to 30.