[IceTV] Make genre fetch more robust
Encode category/genre names in UTF-8 (needed for icetv.de data).

Make the fetch of show["category"]["eit"] more robust - set it to
"0" if the "eit" attribute is missing.

When printing debug log error message when there's a failure to
remap a genre id, only show the first failure for each genre name.