[IceTV] Make screens more skinnable
Explicitly set the title for window names - this is the only practical
way to get window titles translated.

Make initialisation of ChoiceBox in IceTVMain use windowTitle= for
OpenViX and OpenATV, and a fallback initialisation for Beyonwiz,
which doesn't support windowTitle=.

Add a skin name for IceTVMain to make it skinnable.

Set skinName in IceTVOldUserSetup & IceTVCreateLogin and  to their
respective superclasses so that they can use the same skins as the
superclasses if they have a screen skin in the current selected
skin.