Fix issue #766: [IceTV] - error message cleanup
[IceTV]

In _logResponseException(), give more user-friendly error messages
for ConnectionError and for other RequestExceptions by extracting
the underlying error message for ConnectionError and displaying the
IceTV JSON error information where it exists in other RequestExceptions.

Log both the user-friendly error message (where it exists) and the
stringified Exception to the IceTV and debug logs.

Bump version number.